+++ CHANGES.rst
@@ -17,7 +17,7 @@ Bug fixes:
 - *add item here*
 
 
-0.0.2 (in progress)
+0.0.2 (2018-03-05)
 ------------------
 
 Breaking changes:
@@ -26,11 +26,13 @@ Breaking changes:
 
 New features:
 
-- Stress in zone
+- to_list and to_json methods to dump flattened matches
+- distance filter by center distance with +/- tolerance
+- distance filter by range (start, end)
 
 Bug fixes:
 
-- *add item here*
+- distance filter
 
 
 0.0.1 (2018-03-04)